+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Hero2 component.
 - Footer1 component.
 - Footer2 component.
+- Features component.
 - Plume Icon.
 
 ## [0.1.0] - 2026-01-26