+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Footer1 component.
 - Footer2 component.
 - Features component.
+- CallToAction component.
 - Plume Icon.
 
 ## [0.1.0] - 2026-01-26