+++ CHANGELOG.md
@@ -8,8 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- Hero component.
-- Footer component.
+- Hero1 component.
+- Hero2 component.
+- Footer1 component.
+- Footer2 component.
+- Features component.
+- CallToAction component.
 - Plume Icon.
 
 ## [0.1.0] - 2026-01-26