+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.2.0] - 2026-02-27
+
 ### Added
 
 - Block components: Hero1, Hero2, Footer1, Footer2, Features, CallToAction.